requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
SQLAlchemy
flask_bcrypt
Flask-Migrate
python-dotenv
MySQL-python
flask-mysql